ml-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-logging==0.42b0
prometheus-client==0.19.0
structlog==23.2.0
setuptools>=65.0.0
wheel>=0.38.0